feat(ws): add wide and chunked preview wire frames
Preserve every legacy binary layout for u16-sized publications while
selecting additive u32 layouts for wider geometry. Oversized publications
now share a bounded, byte-accounted chunk envelope and strict reassembler.

Teach the TUI to retain u32 source geometry and reassemble chunked canvas
publications with explicit memory budgets. Record the new tags in the
canonical manifest and regenerate Python constants.

+++ crates/hypercolor-tui/src/state.rs
@@ -275,8 +275,8 @@ impl PreviewSource {
 pub struct CanvasFrame {
     pub frame_number: u32,
     pub timestamp_ms: u32,
-    pub width: u16,
-    pub height: u16,
+    pub width: u32,
+    pub height: u32,
     /// RGB pixel data, 3 bytes per pixel, row-major.
     pub pixels: Bytes,
 }
@@ -285,8 +285,8 @@ pub struct CanvasFrame {
 pub struct CanvasPreviewState {
     pub frame_number: u32,
     pub timestamp_ms: u32,
-    pub width: u16,
-    pub height: u16,
+    pub width: u32,
+    pub height: u32,
 }
 
 impl From<&CanvasFrame> for CanvasPreviewState {
